Reformat repository
