--- cfg.py
@@ -1,2 +1,0 @@
-PRINT_INST = True 
-# Prints the currently executed instruction by the CPU to the console
