+++ cli.py
@@ -1,0 +1,4 @@
+from pyGBA.__main__ import main
+
+if __name__ == "__main__":
+    main()
