--- main.py
@@ -1,7 +1,0 @@
-
-def main():
-    print("test")
-
-
-if __name__ == "__main__":
-    main()
