+++ pyGBA/__main__.py
@@ -1,0 +1,4 @@
+from . import *
+
+def main():
+    pass
