+++ pyGBA/core/__init__.py
@@ -1,0 +1,4 @@
+from .cfg import *
+from .cpu import *
+from .mem import *
+from .emu import *
