+++ pyGBA/core/cfg.py
@@ -1,0 +1,2 @@
+# Prints the currently executed instruction by the CPU to the console
+PRINT_INST = True
